refactor(server): carve padi convergence declaration out of padiBinding.ts
W4 ledger L6, sub-item (a). The convergence pre-flight is a fourth concern in
padiBinding.ts, varying with a different volatility than the binder proper
(daemon-lifecycle contract-skew policy, not spawn/supervise/re-serve). Since L3
(#1674) moved the mechanism/ordering/fence into @kolu/surface-daemon-supervisor,
what remains here is padi's thin declaration into that kit — the policy, the
frozen-control-core probe, and the drain plumbing the probe and the "restart"
verb share.

Pure motion: extract PADI_CONVERGENCE_POLICY, probePadiForConvergence,
drainViaControlCore (+ DrainableConn, DRAIN_TEARDOWN_CEILING_MS) into a new
padiConvergence.ts; padiBinding.ts imports back the three it uses. No logic
change. The remote arm (remotePadiBinding.ts) is untouched — it references these
symbols only in prose, never imports them (that's L23's territory).
